backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
# openai  # Uncomment if/when using OpenAI API
